autorisoinnin lisäystä

+++ application/products/views.py
@@ -5,11 +5,12 @@ from application import app, db
 from application.products.models import Product
 from application.products.forms import ProductForm
 from application.offer.models import Offer
+from application.auth.models import User
 
 @app.route("/products", methods=["GET"])
 def products_index():
    
-    return render_template("products/list.html", products = Product.query.all(), offers = Offer.query.all())
+    return render_template("products/list.html", products = Product.query.all(), get_account = User.query.get)
 
 @app.route("/products/new/")
 @login_required
@@ -32,33 +33,51 @@ def products_create():
 
     return redirect(url_for("products_index"))
 
-@app.route("/products/change/price/<product_id>/", methods=["POST"])
-@login_required
-def products_set_new_price(product_id):
+@app.route("/products/<product_id>/", methods=["GET"])
+def product_index(product_id):
     p = Product.query.get(product_id)
-    p.price = request.form.get("new_price")
-    db.session().commit()
+    owner = User.query.get(p.account_id)
+    return render_template("products/product.html", product = p, get_account = User.query.get, get_offers = Offer.by_product, owner = owner)
 
-    return redirect(url_for("products_index"))
+@app.route("/products/change/description/<product_id>/", methods=["POST"])
+@login_required
+def products_set_new_description(product_id):
+    p = Product.query.get(product_id)
+
+    if current_user.id == p.account_id:
+        p.description = request.form.get("new_description")
+        db.session().commit()
+
+    return redirect(url_for("product_index", product_id = product_id))
 
 @app.route("/products/delete/<product_id>/", methods=["POST"])
 @login_required
 def products_delete(product_id):
     p = Product.query.get(product_id)
-    db.session.delete(p)
-    db.session.commit()
 
-    return redirect(url_for("products_index"))
+    if current_user.id == p.account_id:
+        db.session.delete(p)
+        db.session.commit()
+
+    return redirect(url_for("product_index", product_id = product_id))
 
 @app.route("/products/offer/<product_id>", methods=["POST"])
 @login_required
 def products_offer(product_id):
-    
-    o = Offer(request.form.get("offer"))
-    
-    o.account_id = current_user.id
-    o.product_id = product_id
-    db.session().add(o)
-    db.session().commit()
+    amount = request.form.get("offer")
+    p = Product.query.get(product_id)
 
-    return redirect(url_for("products_index"))
+    if current_user.id != p.account_id:
+    
+        if  int(amount) <= p.price:
+            return render_template("products/product.html", e = 1, product = p, get_account = User.query.get, get_offers = Offer.by_product)
+    
+        o = Offer(amount)
+        o.account_id = current_user.id
+        o.product_id = product_id
+        p.price = amount
+        db.session().add(o)
+        db.session().add(p)
+        db.session().commit()
+
+    return redirect(url_for("product_index", product_id = product_id))
